pyproject.toml
[project]
name = "ai-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "youtube-transcript-api>=1.2.3",
]
